chnagin type for DB

+++ CryptoApi/Models/Sentiment.cs
@@ -8,9 +8,9 @@ namespace CryptoApi.Models
         public string id { get; set; }
         public string Symbol { get; set; }
         public DateTime Date { get; set; }
-        public decimal? PositiveReddit { get; set; }
-        public decimal? NeutralReddit { get; set; }
-        public decimal? NegativeReddit { get; set; }
+        public double? PositiveReddit { get; set; }
+        public double? NeutralReddit { get; set; }
+        public double? NegativeReddit { get; set; }
         public int? PostCountReddit { get; set; }
     }
 }
